// src/utills/api/types.ts
// 登录接口
export interface login {
    username: string;
    password: string;
    tenant_code: string;
    verification_code?: number;
}

// Generated by https://quicktype.io

export interface UserInfo {
    auid:          number;
    uid:           number;
    beid:          number;
    ptyid:         number;
    username:      string;
    mobile:        string;
    nickname:      string;
    openid:        string;
    avator:        string;
    access_token:  string;
    access_expire: number;
    refresh_after: number;
}